requirements.txt
numpy==1.26.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.3.2
xgboost==2.0.3

pyspark==3.5.1
findspark==2.0.1

kafka-python==2.0.2

nltk==3.8.1

python-dotenv==1.0.1

streamlit==1.31.1

flask-session==0.5.0

praw==7.7.1
packaging==23.2

snowflake-connector-python==3.0.3